requirements.txt
openai>=1.40.0
python-pptx>=0.6.23
pydantic>=2.7.0
PyYAML>=6.0.1